extracted fetch/clean/data processing separate from Detail component which calls fetchAndCleanArtPost(id) to get post
